Add GPIO_HAL.h
GPIO_HAL is an easy to use HAL api for GPIO pins. It uses the GPIO
structs directly, avoiding the requirement of including chip.h.

Only the code that actually stores the GPIO structs itself needs
GPIO_HAL_LPC.h, which does depend on chip.h.

This has the advantage that most code does not directly depend on chip.h
and is easier to test, for example in a unit-test.