boardconfig: add ADCConfig, GPIO option DIR_INVALID
GPIO pins can be configured with pseudo-direction DIR_INVALID.
When such a pin is accessed, an assertion failure is triggered. This
feature is used for firmware supporting multiple different boards (e.g.
some boards don't have all the GPIO pins).

ADC channels are now also in the BoardConfig, the API is similar to that
for GPIOs.